Move reporting classes to `main`
