+++ community/cypher/compatibility-suite/README.txt
@@ -17,7 +17,7 @@ Main Class:	cucumber.api.cli.Main
 Glue:	cypher
 Feature or folder path:	/path/to/neo4j/checkout/community/cypher/compatibility-suite/src/test/resources/cypher/
 VM options:
-Program arguments:	 --plugin org.jetbrains.plugins.cucumber.java.run.CucumberJvmSMFormatter --plugin cypher.cucumber.db.DatabaseProvider:target/dbs --plugin cypher.cucumber.db.DatabaseConfigProvider:/cypher/db/config/${YOUR_CONFIG_HERE}.json  --plugin cypher.cucumber.reporter.CypherResultReporter:target/reporter-output --monochrome
+Program arguments:	 --plugin org.jetbrains.plugins.cucumber.java.run.CucumberJvmSMFormatter --plugin cypher.cucumber.db.DatabaseProvider:target/dbs --plugin cypher.cucumber.db.DatabaseConfigProvider:/cypher/db/config/${YOUR_CONFIG_HERE}.json  --plugin cypher.feature.reporting.CypherResultReporter:target/reporter-output --monochrome
 Working directory:
 Enviroment variables:
 Use classpath of module:	neo4j-cypher-compatibility-suite
