+++ community/cypher/compatibility-suite/src/main/scala/cypher/feature/reporting/CypherResultReporter.scala
@@ -17,13 +17,12 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package cypher.cucumber.reporter
+package cypher.feature.reporting
 
 import java.io.{File, PrintStream}
 
 import cypher.cucumber.CucumberAdapter
 import cypher.feature.parser.reporting.ChartWriter
-import cypher.feature.steps.CypherTCKSteps
 import gherkin.formatter.model.{Match, Result, Step}
 
 import scala.util.matching.Regex
@@ -36,7 +35,8 @@ object CypherResultReporter {
   }
 }
 
-class CypherResultReporter(producer: OutputProducer, jsonWriter: PrintStream, chartWriter: ChartWriter) extends CucumberAdapter {
+class CypherResultReporter(producer: OutputProducer, jsonWriter: PrintStream, chartWriter: ChartWriter)
+  extends CucumberAdapter {
 
   def this(reportDir: File) = {
     this(producer = JsonProducer,
@@ -46,8 +46,9 @@ class CypherResultReporter(producer: OutputProducer, jsonWriter: PrintStream, ch
 
   private var query: String = null
   private var status: String = Result.PASSED
-  private val pattern: Regex = """running( parametrized)?: (.*)""".r
-  private val newPattern: Regex = CypherTCKSteps.EXECUTING_QUERY.r
+  // This is copied from the constant CypherTCKSteps.EXECUTING_QUERY
+  // Perhaps CypherTCKSteps will move to `main` from `test` some day
+  private val queryPattern: Regex = "^executing query: (.*)$".r
 
   override def done(): Unit = {
     jsonWriter.println(producer.dump())
@@ -62,8 +63,7 @@ class CypherResultReporter(producer: OutputProducer, jsonWriter: PrintStream, ch
   override def step(step: Step) {
     if (step.getKeyword.trim == "When") {
       step.getName match {
-        case pattern(_, q) => query = q
-        case newPattern(q) => query = q
+        case queryPattern(q) => query = q
       }
     }
   }
